refactor(questionnaires): split questionnaires/new page under the line cap (#544)
Extract the create page's two settings panels into co-located components:
`QuestionnaireCreateBasicInfo` (name/description/type/requires-evaluation/
min-score/evaluation-mode, with the type & evaluation label maps that only
they use) and `QuestionnaireCreateAdvancedSettings` (shuffle, members-exempt,
per-event, LLM guidelines, durations, max-attempts). Both take `$bindable`
props for the fields the panels edit; read-only derived (effective/eval flags,
LLM warning) are passed down. Pure markup move — no behaviour change.

`new/+page.svelte`: 989 -> 490 lines.
